requirements.txt
pip~=23.2.1
asgiref~=3.8.1
sqlparse~=0.5.1
django==4.1
flake8==5.0.4
flake8-quotes==3.3.1
flake8-variables-names==0.0.5
pep8-naming==0.13.2
django-debug-toolbar==3.4.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.22.1
Pillow==9.1.1
setuptools==75.2.0

